refactoring: useAlbums hook
