+++ src/components/burden/Burden.tsx
@@ -1,32 +1,11 @@
-import axios from 'axios';
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 
-import { Album, AlbumPhoto } from '../../types/api/album';
+import { useAlbums } from './hooks/useAlbums';
 
 const Burden = () => {
-  const [albums, setAlbums] = useState<Album[]>([]);
-  const [loading, setLoading] = useState<boolean>(false);
+  const { getAlbums, loading, albums } = useAlbums();
 
-  useEffect(() => {
-    setLoading(true);
-
-    const fetchAlbums = async () => {
-      const url = 'https://jsonplaceholder.typicode.com/albums';
-      const albums = (await axios.get<Album[]>(url)).data.slice(0, 2);
-
-      for (const album of albums) {
-        const url = `https://jsonplaceholder.typicode.com/albums/${album.id}/photos`;
-        const photos = (await axios.get<AlbumPhoto[]>(url)).data.slice(0, 3);
-        console.log(photos);
-        album.photos = photos;
-      }
-
-      setAlbums(albums);
-      setLoading(false);
-    };
-
-    fetchAlbums();
-  }, []);
+  useEffect(() => getAlbums(), []);
 
   return (
     <>
@@ -47,7 +26,13 @@ const Burden = () => {
                   <td>{album.title}</td>
                   <td>
                     {album.photos.map((photo) => {
-                      return <img width={32} src={photo.thumbnailUrl} />;
+                      return (
+                        <img
+                          key={photo.id}
+                          width={32}
+                          src={photo.thumbnailUrl}
+                        />
+                      );
                     })}
                   </td>
                 </tr>
